requirements.txt
einops==0.4.1
h5py==2.10.0
hydra-core==1.3.2
librosa==0.9.1
matplotlib==3.3.3
moviepy==1.0.3
numba==0.52.0
numpy==1.24.2
omegaconf==2.0.6
pandas==1.3.5
positional_encodings==6.0.1
pypesq==1.2.4
python_hostlist==1.23.0
scikit_learn==1.2.1
scipy==1.7.3
seaborn==0.11.1
sounddevice==0.4.1
tabulate==0.8.10
timm==0.6.11
torch==1.11.0
tqdm==4.64.1
vector_quantize_pytorch==0.10.10
